+++ src/python/CRABClient/Commands/getlog.py
@@ -1,3 +1,6 @@
+from __future__ import print_function
+from __future__ import division
+
 import CRABClient.Emulator
 from CRABClient import __version__
 from CRABClient.ClientUtilities import colors
@@ -10,6 +13,13 @@ from ServerUtilities import getProxiedWebDir
 
 class getlog(getcommand):
     """
+    Important: code here is identical to the old getlog implementation (aside from setting the subresource to
+    'logs2' when calling getcommand and the names of the command/class themselves). This was done because trying to
+    avoid copy-paste code isn't worth the effort in this case. When the status2 is working correctly, old code will
+    be easily removed and replaced with the 'getlog2' version. Also, the command 'getlog' itself is deprecated and
+    we don't expect to make any changes to it until it's removed.
+
+    Class description:
     Retrieve the log files of a number of jobs specified by the -q/--quantity option.
     -q logfiles per exit code are returned if transferLogs = False; otherwise all the log files
     collected by the LogCollect job are returned. The task is identified by the -d/--dir option.
@@ -43,7 +53,8 @@ class getlog(getcommand):
                 self.logger.info("%sSuccess%s: All files successfully retrieved." % (colors.GREEN,colors.NORMAL))
             returndict = {'success': success, 'failed': failed}
         else:
-            returndict = getcommand.__call__(self, subresource = 'logs')
+            # Different from the old getlog code: set 'logs2' as subresource so that 'getcommand' uses the new logic.
+            returndict = getcommand.__call__(self, subresource = 'logs2')
             if ('success' in returndict and not returndict['success']) or \
                ('failed'  in returndict and returndict['failed']):
                 msg = "You can use the --short option to retrieve a short version of the log files from the Grid scheduler."
